+++ modules/AWSPowerShell/Cmdlets/AmazonECS/Basic/Remove-ECSAttribute-Cmdlet.cs
@@ -58,7 +58,7 @@ namespace Amazon.PowerShell.Cmdlets.ECS
         /// <summary>
         /// <para>
         /// <para>The short name or full Amazon Resource Name (ARN) of the cluster that contains the
-        /// resource to apply attributes. If you do not specify a cluster, the default cluster
+        /// resource to delete attributes. If you do not specify a cluster, the default cluster
         /// is assumed.</para>
         /// </para>
         /// </summary>